% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AGTI_curso-data.R
\docType{data}
\name{AGTI_curso}
\alias{AGTI_curso}
\title{Aluno de graduação em Tempo Integral por curso, UFERSA (2013-2017)}
\format{Data frame com 45 linhas e 7 colunas.}
\usage{
data(AGTI_curso)
}
\description{
Aluno de graduação em tempo integral, por curso, na Universidade Federal
Rural do Semi-Árido – 2013/2017.
}
\details{
\itemize{
  \item  cursos
  \item  ano 2013 (k13)
  \item  ano 2014 (k14)
  \item  ano 2015 (k15)
  \item  ano 2016 (k16)
  \item  ano 2017 (k17)
  \item  razão 2017/2013 (k17/k13)
}
}
\keyword{datasets}
