% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capes-data.R
\docType{data}
\name{capes}
\alias{capes}
\title{Conceito CAPES/MEC para a pós-graduação (2017)}
\format{Data frame com 17 linhas e 2 colunas.}
\usage{
data(capes)
}
\description{
Conceito capes para programas de pós-graduação da UFERSA - 2017.
}
\details{
\itemize{
  \item  Programa de pós-graduação (curso)
  \item  Conceito Capes (conceito)
}
}
\keyword{datasets}
