% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diplomados-data.R
\docType{data}
\name{diplomados}
\alias{diplomados}
\title{Número de diplomados e ingressantes na graduação - UFERSA}
\format{Data frame com 38 linhas e 11 colunas.}
\usage{
data(diplomados)
}
\description{
A tabela foi construída com base no item 7 do anexo Indicadores do TCU do Relatório
de Gestão 2017 da UFERSA. Os cursos de engenharia não possuem concorrência
disponível, pois os alunos ingressam no curso de Ciência e Tecnologia para
depois, por meio de seleção interna, iniciarem os estudos em alguma área da
engenharia.
}
\details{
\itemize{
  \item  Curso
  \item  Campus onde está localizado o curso - Angicos, Caraúbas, Mossoró ou
  Pau dos Ferros (campus)
  \item Duração esperada do curso, em anos (duracao)
  \item Número de alunos ingressantes em 2017 - duracao. (ingressos)
  \item Nota do curso no Enade informada na página
  do curso no portal da UFERSA (enade)
  \item Área do curso, podendo ser Ciências Exatas e da Terra (exatas),
  Ciências Biológicas (biologicas), Engenharias, Humanas, Agrárias,
  Ciências Sociais Aplicadas,
}
}
\keyword{datasets}
