% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forplad_assist-data.R
\docType{data}
\name{forplad_assist}
\alias{forplad_assist}
\title{Indicadores de Assistência Estudantil (2013-2017) - Forplad}
\format{Data frame com 4 linhas e 6 colunas.}
\usage{
data(forplad_assist)
}
\description{
Indicadores de assistência estudantil da UFERSA do Fórum Nacional de Pró-Reitores de
Planejamento e de Administração das Instituições Federais de Ensino Superior (Forplad).
}
\details{
\itemize{
  \item Número de alunos assistidos pelo Programa
  Nacional de Assistência Estudantil (N_ALUNOS_PNAES)
  \item Índice de evasão de alunos assistidos em relação ao total de alunos
  vinculados ao Programa Nacional de Assistência Estudantil (IEv_pnaes)
  \item Índice de retenção de alunos assistidos em relação ao total de
  alunos da vinculados ao Programa Nacional de Assistência Estudantil (IRt_pnaes)
  \item Índice de alunos atendidos em relação ao total de
  alunos com direito a auxílios (IAAT_da)
}
}
\keyword{datasets}
