% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forplad_ext-data.R
\docType{data}
\name{forplad_ext}
\alias{forplad_ext}
\title{Indicadores de extensão (2013-2017) - Forplad}
\format{Data frame com 5 linhas e 6 colunas.}
\usage{
data(forplad_ext)
}
\description{
Indicadores de extensão da UFERSA do Fórum Nacional de Pró-Reitores de
Planejamento e de Administração das Instituições Federais de Ensino Superior (Forplad).
}
\details{
\itemize{
  \item  Taxa de alunos com bolsas de extensão (TABEx)
  \item  Taxa de alunos Extensionistas (TAExt)
  \item  Taxa de volume de recursos financeiros captados por editais e similares (TRFE)
  \item  Número de servidores membros de ações de extensão (N_SERVIDORES)
  \item  Número de participantes como “atendidos” em ações de extensão (N_ATENDIDOS)
}
}
\keyword{datasets}
