% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forplad_grad-data.R
\docType{data}
\name{forplad_grad}
\alias{forplad_grad}
\title{Indicadores de Graduação (2013-2017) - Forplad}
\format{Data frame com 6 linhas e 6 colunas.}
\usage{
data(forplad_grad)
}
\description{
Indicadores de graduação da UFERSA do Fórum Nacional de Pró-Reitores de
Planejamento e de Administração das Instituições Federais de Ensino Superior (Forplad).
}
\details{
\itemize{
  \item Número de cursos de graduação (NCurGr)
  \item Número de alunos ingressantes na graduação (NTIng)
  \item Número de alunos concluintes na graduação (NCGr)
  \item Índice de evasão dos cursos de graduação (Eva)
  \item Índice de retenção dos cursos de graduação (IRETA)
  \item Índice de desempenho acadêmico dos cursos de graduação (IDesAcadGra)
}
}
\keyword{datasets}
