% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forplad_pesqPos-data.R
\docType{data}
\name{forplad_pesqPos}
\alias{forplad_pesqPos}
\title{Indicadores de pesquisa e pós-graduação (2013-2017) - Forplad}
\format{Data frame com 4 linhas e 6 colunas.}
\usage{
data(forplad_pesqPos)
}
\description{
Indicadores de pesquisa e pós-graduação da UFERSA do Fórum Nacional de Pró-Reitores de
Planejamento e de Administração das Instituições Federais de Ensino Superior (Forplad).
}
\details{
\itemize{
  \item Número de titulados em programas de pós-graduação (NTPPGr)
  \item Número de projetos de pesquisa financiados (NPPqF)
  \item Número de projetos de cooperação internacional aprovados (NPCIAp)
  \item Índice de internacionalização da produção científica (IIPCient)
}
}
\keyword{datasets}
