% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ouvidoria-data.R
\docType{data}
\name{ouvidoria}
\alias{ouvidoria}
\title{Quantitativo de chamadas realizadas junto à Ouvidoria da UFERSA (2013-2017)}
\format{Data frame com 8 linhas e 6 colunas.}
\usage{
data(ouvidoria)
}
\description{
Quantitativo de chamadas realizadas junto à Ouvidoria na UFERSA no período
2013-2017 por tipo de chamada.
}
\details{
\itemize{
  \item  Categoria da chamada, podendo ser denúncia, reclamação,
  crítica, dúvida, solicitação, informação, sugestão ou elogio (categoria)
  \item  ano 2013 (k13)
  \item  ano 2014 (k14)
  \item  ano 2015 (k15)
  \item  ano 2016 (k16)
  \item  ano 2017 (k17)
}
}
\keyword{datasets}
