% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qalunos-data.R
\docType{data}
\name{Qalunos}
\alias{Qalunos}
\title{Quantidade de alunos matriculados por curso, UFERSA (2013-2017)}
\format{Data frame com 46 linhas e 6 colunas.}
\usage{
data(Qalunos)
}
\description{
Quantidade de alunos matriculados nos cursos de graduação –
2013/2017. Os dados são referentes as matriculas de 2016.2 e 2017.1
}
\details{
\itemize{
  \item  cursos
  \item  ano 2013 (k13)
  \item  ano 2014 (k14)
  \item  ano 2015 (k15)
  \item  ano 2016 (k16)
  \item  ano 2017 (k17)
}
}
\keyword{datasets}
