% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qalunos_pos-data.R
\docType{data}
\name{Qalunos_pos}
\alias{Qalunos_pos}
\title{Quantidade de alunos matriculados na pós-graduação (mestrado)
por curso, UFERSA (2013-2017)}
\format{Data frame com 58 linhas e 9 colunas.}
\usage{
data(Qalunos_pos)
}
\description{
Alunos dos programas de pós-graduação Stricto sensu (Mestrado) na
Universidade Federal Rural do Semi-Árido – 2013/2017
Cursos de Pós-Graduação Stricto sensu - Mestrado
}
\details{
\itemize{
  \item  cursos
  \item  ano
  \item  Número de inscritos na seleção do mestrado (inscritos)
  \item  Número de aprovados (aprovados)
  \item  Concorrência (concorrencia)
  \item  Alunos remanescentes do ano anterior (alunos_remanescentes)
  \item  Alunos ingressantes (alunos_novos)
  \item  Total de alunos no mestrado (alunos_total)
  \item  Número de dissertações defendidas (dissertacoes_defendidas)
}
}
\keyword{datasets}
