% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcu_AG-data.R
\docType{data}
\name{tcu_AG}
\alias{tcu_AG}
\title{Número de Alunos da Graduação da UFERSA (2013-2017) - TCU}
\format{Data frame com 46 linhas e 6 colunas.}
\usage{
data(tcu_AG)
}
\description{
Indicadores do número de alunos de graduação, por curso, da UFERSA
para 2016.2 e 2017.1 e a média.
}
\details{
\itemize{
  \item  Número do curso (numero_curso)
  \item  curso
  \item  AG em 2016.2 (k16.2)
  \item  AG em 2017.1 (k17.1)
  \item  Média do período (media)
  \item  Campus onde está localizado o curso - Angicos, Caraúbas, Mossoró ou
  Pau dos Ferros (campus)
}
}
\keyword{datasets}
