% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcu_AGE-data.R
\docType{data}
\name{tcu_AGE}
\alias{tcu_AGE}
\title{Aluno Equivalente da Graduação da UFERSA - TCU}
\format{Data frame com 46 linhas e 6 colunas.}
\usage{
data(tcu_AGE)
}
\description{
Números do indicador Aluno equivalente da graduação, por curso, da UFERSA. Fonte: Relatório
de Gestão 2017.
}
\details{
\itemize{
  \item  Número do curso (numero_curso)
  \item  curso
  \item  Aluno de Graduação em Tempo Integral do curso (AGTI)
  \item  Peso do Grupo no Curso (peso_grupo)
  \item  Aluno Equivalente da Graduação (AGE)
  \item  Campus onde está localizado o curso - Angicos, Caraúbas, Mossoró ou
  Pau dos Ferros (campus)
}
}
\keyword{datasets}
