% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcu_AGTI-data.R
\docType{data}
\name{tcu_AGTI}
\alias{tcu_AGTI}
\title{Aluno da Graduação em Tempo Integral da UFERSA (2016.2-2017.1) - TCU}
\format{Data frame com 46 linhas e 8 colunas.}
\usage{
data(tcu_AGTI)
}
\description{
Números do indicador Aluno equivalente da graduação, por curso, da UFERSA. Fonte: Relatório
de Gestão 2017.
}
\details{
\itemize{
  \item  Número do curso (numero_curso)
  \item  curso
  \item  Número de diplomados (2016.2 + 2017.1) (NDI)
  \item  Duração padrão do curso (DPC)
  \item  Fator de retenção (fator_retencao)
  \item  Número de Ingressantes (NI)
  \item  Aluno de Graduação em Tempo Integral do curso (AGTI)
  \item  Campus onde está localizado o curso - Angicos, Caraúbas, Mossoró ou
  Pau dos Ferros (campus)
}
}
\keyword{datasets}
