% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcu_APG-data.R
\docType{data}
\name{TCU_APG}
\alias{TCU_APG}
\alias{tcu_APG}
\title{Número de Alunos da Pós-Graduação da UFERSA (2017.1-2017.2) - TCU}
\format{Data frame com 17 linhas e 5 colunas.}
\usage{
data(TCU_APG)
}
\description{
Número de Alunos da Pós-Graduação, por curso, da UFERSA. Fonte: Relatório
de Gestão 2017.
}
\details{
\itemize{
  \item  Número do curso (numero_curso)
  \item  curso
  \item  Primeiro semestre 2017 (k17.1)
  \item  Segundo semestre 2017 (k17.2)
  \item  Média anual (media)
}
}
\keyword{datasets}
