% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcu_gestao-data.R
\docType{data}
\name{tcu_gestao}
\alias{tcu_gestao}
\title{Número de Alunos da Pós-Graduação da UFERSA (2013-2017) - TCU}
\format{Data frame com 9 linhas e 6 colunas.}
\usage{
data(tcu_gestao)
}
\description{
Indicadores de Gestão, UFERSA – 2013/2017, conforme Decisão nº 408/2002 – TCU – Plenário
}
\details{
\itemize{
  \item  2013 (k13)
  \item  2014 (k14)
  \item  2015 (k15)
  \item  2016 (k16)
  \item  2017 (k17)
  \item  Razão 2017/2013 (k17/k13)
}
}
\keyword{datasets}
