% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcu_NFE-data.R
\docType{data}
\name{tcu_NFE}
\alias{tcu_NFE}
\title{Número de Funcionários Equivalentes da UFERSA (2017) - TCU}
\format{Data frame com 3 linhas e 6 colunas.}
\usage{
data(tcu_NFE)
}
\description{
Número de funcionários equivalentes, UFERSA - 2017.
}
\details{
\itemize{
  \item  regime
  \item  tecnicos_administrativos
  \item  terceirizados
  \item  afastados
  \item  total
  \item  funcionario_equivalente
}
}
\keyword{datasets}
