% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcu_NPE-data.R
\docType{data}
\name{tcu_NPE}
\alias{tcu_NPE}
\title{Número de Professores Equivalentes da UFERSA (2017) - TCU}
\format{Data frame com 3 linhas e 6 colunas.}
\usage{
data(tcu_NPE)
}
\description{
Número de Professores equivalentes, UFERSA - 2017.
}
\details{
\itemize{
  \item  regime
  \item  efetivo
  \item  substituto_ou_visitante
  \item  afastados
  \item  total
  \item  funcionario_equivalente
}
}
\keyword{datasets}
