% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcu_primario-data.R
\docType{data}
\name{tcu_primario}
\alias{tcu_primario}
\title{Número de Professores Equivalentes da UFERSA (2017) - TCU}
\format{Data frame com 8 linhas e 5 colunas.}
\usage{
data(tcu_primario)
}
\description{
Indicadores primários utilizados para cálculo dos indicadores de gestão da UFERSA,
conforme Decisão TCU 408/2002 - Exercício 2013/2017.
Custo corrente sem HU, Número de Professor Equivalente,
Número de Funcionários Equivalente sem HU, Total de Alunos Regularmente Matriculados
na Graduação, Aluno de Pós-Graduação, Aluno Equivalente de Graduação,
Aluno de Graduação em Tempo Integral, Aluno de Pós-Graduação em Tempo Integral.
}
\details{
\itemize{
  \item  2013 (k13)
  \item  2014 (k14)
  \item  2015 (k15)
  \item  2016 (k16)
  \item  2017 (k17)
}
}
\keyword{datasets}
